Support flat container icon URL in Azure Search (#561)
Progress on https://github.com/NuGet/NuGetGallery/issues/7261
