+++ src/NuGet.Services.AzureSearch/AzureSearchJobConfiguration.cs
@@ -14,12 +14,19 @@ namespace NuGet.Services.AzureSearch
         public string StorageContainer { get; set; }
         public string StoragePath { get; set; }
         public string GalleryBaseUrl { get; set; }
+        public string FlatContainerBaseUrl { get; set; }
+        public string FlatContainerContainerName { get; set; }
 
         public AzureSearchScoringConfiguration Scoring { get; set; }
 
         public Uri ParseGalleryBaseUrl()
         {
             return new Uri(GalleryBaseUrl, UriKind.Absolute);
+        }
+
+        public Uri ParseFlatContainerBaseUrl()
+        {
+            return new Uri(FlatContainerBaseUrl, UriKind.Absolute);
         }
 
         public string NormalizeStoragePath()
