+++ src/NuGet.Services.AzureSearch/HijackDocumentBuilder.cs
@@ -2,7 +2,6 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using Microsoft.Extensions.Options;
 using NuGet.Protocol.Catalog;
 using NuGet.Services.Entities;
 
@@ -10,11 +9,11 @@ namespace NuGet.Services.AzureSearch
 {
     public class HijackDocumentBuilder : IHijackDocumentBuilder
     {
-        private readonly IOptionsSnapshot<AzureSearchJobConfiguration> _options;
+        private readonly IBaseDocumentBuilder _baseDocumentBuilder;
 
-        public HijackDocumentBuilder(IOptionsSnapshot<AzureSearchJobConfiguration> options)
+        public HijackDocumentBuilder(IBaseDocumentBuilder baseDocumentBuilder)
         {
-            _options = options ?? throw new ArgumentNullException(nameof(options));
+            _baseDocumentBuilder = baseDocumentBuilder ?? throw new ArgumentNullException(nameof(baseDocumentBuilder));
         }
 
         public KeyedDocument Keyed(
@@ -64,7 +63,7 @@ namespace NuGet.Services.AzureSearch
                 lastCommitTimestamp: leaf.CommitTimestamp,
                 lastCommitId: leaf.CommitId,
                 changes: changes);
-            DocumentUtilities.PopulateMetadata(document, normalizedVersion, leaf, _options.Value.ParseGalleryBaseUrl());
+            _baseDocumentBuilder.PopulateMetadata(document, normalizedVersion, leaf);
             document.Listed = leaf.IsListed();
 
             return document;
@@ -85,13 +84,13 @@ namespace NuGet.Services.AzureSearch
                 lastCommitId: null,
                 normalizedVersion: package.NormalizedVersion,
                 changes: changes);
-            DocumentUtilities.PopulateMetadata(document, packageId, package, _options.Value.ParseGalleryBaseUrl());
+            _baseDocumentBuilder.PopulateMetadata(document, packageId, package);
             document.Listed = package.Listed;
 
             return document;
         }
 
-        private static void PopulateLatest<T>(
+        private void PopulateLatest<T>(
             T document,
             string packageId,
             string normalizedVersion,
@@ -101,7 +100,7 @@ namespace NuGet.Services.AzureSearch
             HijackDocumentChanges changes) where T : KeyedDocument, HijackDocument.ILatest
         {
             PopulateKey(document, packageId, normalizedVersion);
-            DocumentUtilities.PopulateCommitted(
+            _baseDocumentBuilder.PopulateCommitted(
                 document,
                 lastUpdatedFromCatalog,
                 lastCommitTimestamp,
