+++ src/NuGet.Services.AzureSearch/NuGet.Services.AzureSearch.csproj
@@ -48,8 +48,10 @@
     <Compile Include="Analysis\PackageIdCustomAnalyzer.cs" />
     <Compile Include="Analysis\PackageIdCustomTokenizer.cs" />
     <Compile Include="AzureSearchScoringConfiguration.cs" />
+    <Compile Include="BaseDocumentBuilder.cs" />
     <Compile Include="DatabaseOwnerFetcher.cs" />
     <Compile Include="IAzureSearchTelemetryService.cs" />
+    <Compile Include="IBaseDocumentBuilder.cs" />
     <Compile Include="IDatabaseOwnerFetcher.cs" />
     <Compile Include="Owners2AzureSearch\IOwnerIndexActionBuilder.cs" />
     <Compile Include="Owners2AzureSearch\IOwnerSetComparer.cs" />
