+++ src/NuGet.Services.AzureSearch/SearchDocumentBuilder.cs
@@ -10,10 +10,14 @@ namespace NuGet.Services.AzureSearch
 {
     public class SearchDocumentBuilder : ISearchDocumentBuilder
     {
+        private readonly IBaseDocumentBuilder _baseDocumentBuilder;
         private readonly IOptionsSnapshot<AzureSearchJobConfiguration> _options;
 
-        public SearchDocumentBuilder(IOptionsSnapshot<AzureSearchJobConfiguration> options)
+        public SearchDocumentBuilder(
+            IBaseDocumentBuilder baseDocumentBuilder,
+            IOptionsSnapshot<AzureSearchJobConfiguration> options)
         {
+            _baseDocumentBuilder = baseDocumentBuilder ?? throw new ArgumentNullException(nameof(baseDocumentBuilder));
             _options = options ?? throw new ArgumentNullException(nameof(options));
         }
 
@@ -177,7 +181,7 @@ namespace NuGet.Services.AzureSearch
                 isLatest: isLatest,
                 fullVersion: fullVersion,
                 owners: owners);
-            DocumentUtilities.PopulateMetadata(document, normalizedVersion, leaf, _options.Value.ParseGalleryBaseUrl());
+            _baseDocumentBuilder.PopulateMetadata(document, normalizedVersion, leaf);
 
             return document;
         }
@@ -207,14 +211,14 @@ namespace NuGet.Services.AzureSearch
                 isLatest: isLatest,
                 fullVersion: fullVersion,
                 owners: owners);
-            DocumentUtilities.PopulateMetadata(document, packageId, package, _options.Value.ParseGalleryBaseUrl());
+            _baseDocumentBuilder.PopulateMetadata(document, packageId, package);
             document.TotalDownloadCount = totalDownloadCount;
             document.LogOfDownloadCount = Math.Log(totalDownloadCount, _options.Value.Scoring.DownloadCountLogBase);
 
             return document;
         }
 
-        private static void PopulateVersions<T>(
+        private void PopulateVersions<T>(
             T document,
             string packageId,
             SearchFilters searchFilters,
@@ -226,7 +230,7 @@ namespace NuGet.Services.AzureSearch
             bool isLatest) where T : KeyedDocument, SearchDocument.IVersions
         {
             PopulateKey(document, packageId, searchFilters);
-            DocumentUtilities.PopulateCommitted(
+            _baseDocumentBuilder.PopulateCommitted(
                 document,
                 lastUpdatedFromCatalog,
                 lastCommitTimestamp,
@@ -241,7 +245,7 @@ namespace NuGet.Services.AzureSearch
             document.Key = DocumentUtilities.GetSearchDocumentKey(packageId, searchFilters);
         }
 
-        private static void PopulateUpdateLatest(
+        private void PopulateUpdateLatest(
             SearchDocument.UpdateLatest document,
             string packageId,
             SearchFilters searchFilters,
@@ -286,7 +290,7 @@ namespace NuGet.Services.AzureSearch
             var document = new SearchDocument.UpdateOwners();
 
             PopulateKey(document, packageId, searchFilters);
-            DocumentUtilities.PopulateCommitted(
+            _baseDocumentBuilder.PopulateCommitted(
                 document,
                 lastUpdatedFromCatalog: false,
                 lastCommitTimestamp: null,
