+++ tests/NuGet.Services.AzureSearch.Tests/DocumentUtilitiesFacts.cs
@@ -2,18 +2,13 @@
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
 using System;
-using System.Collections.Generic;
 using System.Linq;
-using NuGet.Protocol.Catalog;
-using NuGet.Services.AzureSearch.Support;
 using Xunit;
 
 namespace NuGet.Services.AzureSearch
 {
     public class DocumentUtilitiesFacts
     {
-        private static readonly Uri GalleryBaseUrl = new Uri(Data.GalleryBaseUrl, UriKind.Absolute);
-
         public class GetHijackDocumentKey
         {
             [Theory]
@@ -62,158 +57,6 @@ namespace NuGet.Services.AzureSearch
 
                     Assert.Equal(expected + "-" + searchFilters, actual);
                 }
-            }
-        }
-
-        public class PopulateMetadataWithCatalogLeaf
-        {
-            private const string NormalizedVersion = "1.0.0";
-
-            [Theory]
-            [InlineData("any")]
-            [InlineData("agnostic")]
-            [InlineData("unsupported")]
-            [InlineData("fakeframework1.0")]
-            public void DoesNotIncludeDependencyVersionSpecialFrameworks(string framework)
-            {
-                var leaf = new PackageDetailsCatalogLeaf
-                {
-                    PackageId = Data.PackageId,
-                    PackageVersion = NormalizedVersion,
-                    DependencyGroups = new List<PackageDependencyGroup>
-                    {
-                        new PackageDependencyGroup
-                        {
-                            TargetFramework = framework,
-                            Dependencies = new List<PackageDependency>
-                            {
-                                new PackageDependency
-                                {
-                                    Id = "NuGet.Versioning",
-                                    Range = "2.0.0",
-                                },
-                                new PackageDependency
-                                {
-                                    Id = "NuGet.Frameworks",
-                                    Range = "3.0.0",
-                                },
-                            },
-                        },
-                    },
-                };
-                var full = new HijackDocument.Full();
-
-                DocumentUtilities.PopulateMetadata(full, NormalizedVersion, leaf, GalleryBaseUrl);
-
-                Assert.Equal("NuGet.Versioning:2.0.0|NuGet.Frameworks:3.0.0", full.FlattenedDependencies);
-            }
-
-            [Theory]
-            [InlineData("any", ":")]
-            [InlineData("net40", "::net40")]
-            [InlineData("NET45", "::net45")]
-            [InlineData(".NETFramework,Version=4.0", "::net40")]
-            public void AddsEmptyDependencyGroup(string framework, string expected)
-            {
-                var leaf = new PackageDetailsCatalogLeaf
-                {
-                    PackageId = Data.PackageId,
-                    PackageVersion = NormalizedVersion,
-                    DependencyGroups = new List<PackageDependencyGroup>
-                    {
-                        new PackageDependencyGroup
-                        {
-                            TargetFramework = framework,
-                        },
-                    },
-                };
-                var full = new HijackDocument.Full();
-
-                DocumentUtilities.PopulateMetadata(full, NormalizedVersion, leaf, GalleryBaseUrl);
-
-                Assert.Equal(expected, full.FlattenedDependencies);
-            }
-
-            [Fact]
-            public void AddsEmptyStringForDependencyVersionAllRange()
-            {
-                var leaf = new PackageDetailsCatalogLeaf
-                {
-                    PackageId = Data.PackageId,
-                    PackageVersion = NormalizedVersion,
-                    DependencyGroups = new List<PackageDependencyGroup>
-                    {
-                        new PackageDependencyGroup
-                        {
-                            TargetFramework = "net40",
-                            Dependencies = new List<PackageDependency>
-                            {
-                                new PackageDependency
-                                {
-                                    Id = "NuGet.Versioning",
-                                    Range = "(, )"
-                                }
-                            },
-                        },
-                    },
-                };
-                var full = new HijackDocument.Full();
-
-                DocumentUtilities.PopulateMetadata(full, NormalizedVersion, leaf, GalleryBaseUrl);
-
-                Assert.Equal("NuGet.Versioning::net40", full.FlattenedDependencies);
-            }
-
-            [Theory]
-            [InlineData("1.0.0", "1.0.0")]
-            [InlineData("1.0.0-beta+git", "1.0.0-beta")]
-            [InlineData("[1.0.0-beta+git, )", "1.0.0-beta")]
-            [InlineData("[1.0.0, 1.0.0]", "[1.0.0]")]
-            [InlineData("[1.0.0, 2.0.0)", "[1.0.0, 2.0.0)")]
-            [InlineData("[1.0.0, )", "1.0.0")]
-            public void AddShortFormOfDependencyVersionRange(string input, string expected)
-            {
-                var leaf = new PackageDetailsCatalogLeaf
-                {
-                    PackageId = Data.PackageId,
-                    PackageVersion = NormalizedVersion,
-                    DependencyGroups = new List<PackageDependencyGroup>
-                    {
-                        new PackageDependencyGroup
-                        {
-                            TargetFramework = "net40",
-                            Dependencies = new List<PackageDependency>
-                            {
-                                new PackageDependency
-                                {
-                                    Id = "NuGet.Versioning",
-                                    Range = input
-                                }
-                            },
-                        },
-                    },
-                };
-                var full = new HijackDocument.Full();
-
-                DocumentUtilities.PopulateMetadata(full, NormalizedVersion, leaf, GalleryBaseUrl);
-
-                Assert.Equal("NuGet.Versioning:" + expected + ":net40", full.FlattenedDependencies);
-            }
-
-            [Fact]
-            public void AllowNullDependencyGroups()
-            {
-                var leaf = new PackageDetailsCatalogLeaf
-                {
-                    PackageId = Data.PackageId,
-                    PackageVersion = NormalizedVersion,
-                    DependencyGroups = null
-                };
-                var full = new HijackDocument.Full();
-
-                DocumentUtilities.PopulateMetadata(full, NormalizedVersion, leaf, GalleryBaseUrl);
-
-                Assert.Null(full.FlattenedDependencies);
             }
         }
     }
